requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
pandas==2.2.2
tqdm==4.66.4
PyPDF2==3.0.1
sentence-transformers==2.6.1
numpy==1.26.4
requests==2.31.0
python-dateutil==2.8.2
